+++ packages/components/src/components/bal-footer/bal-footer.tsx
@@ -7,6 +7,7 @@ import {
   BalLanguage,
   detachComponentToConfig,
   attachComponentToConfig,
+  updateBalLanguage,
 } from '../../config'
 
 @Component({
@@ -15,6 +16,7 @@ import {
 export class Footer implements BalConfigObserver {
   @State() links: FooterLink[] = []
   @State() language: BalLanguage = defaultConfig.language
+  @State() allowedLanguages: BalLanguage[] = defaultConfig.allowedLanguages
 
   /**
    * @deprecated The languages in which the links will appear.
@@ -22,9 +24,14 @@ export class Footer implements BalConfigObserver {
   @Prop() locale: 'en' | 'de' | 'fr' | 'it' | '' = ''
 
   /**
-   * If `true` the default Baloise links will be hidden.
+   * If `true` the legal Baloise links will be hidden.
    */
   @Prop() hideLinks = false
+
+  /**
+   * If `true` the language selection will be hidden.
+   */
+  @Prop() hideLanguageSelection = false
 
   disconnectedCallback() {
     detachComponentToConfig(this)
@@ -37,6 +44,7 @@ export class Footer implements BalConfigObserver {
 
   configChanged(config: BalConfigState) {
     this.language = config.language
+    this.allowedLanguages = config.allowedLanguages
     this.updateFooterLinks()
   }
 
@@ -46,6 +54,10 @@ export class Footer implements BalConfigObserver {
       this.language = this.locale
       this.updateFooterLinks()
     }
+  }
+
+  changeLanguage(language: BalLanguage) {
+    updateBalLanguage(language)
   }
 
   updateFooterLinks() {
@@ -68,12 +80,32 @@ export class Footer implements BalConfigObserver {
           }}
         >
           <slot></slot>
-          <div class="footer-links-container p-1" style={{ display: this.hideLinks ? 'none' : 'block' }}>
-            {this.links.map(link => (
-              <a class="is-link is-inverted pr-4" href={link.link}>
-                {link.label}
-              </a>
-            ))}
+          <div class="container footer-links-container">
+            <div class="legal-links pt-4" style={{ display: this.hideLinks ? 'none' : 'flex' }}>
+              {this.links.map(link => (
+                <a class="is-link is-inverted pr-4" href={link.link}>
+                  {link.label}
+                </a>
+              ))}
+            </div>
+            <div
+              class="language-links pt-4"
+              style={{ display: this.hideLanguageSelection || this.allowedLanguages.length <= 1 ? 'none' : 'flex' }}
+            >
+              {this.allowedLanguages.map(lang => (
+                <a
+                  class={[
+                    'is-link',
+                    'is-inverted',
+                    'pr-4',
+                    this.language.toLowerCase() == lang.toLowerCase() ? 'is-current' : '',
+                  ].join(' ')}
+                  onClick={() => this.changeLanguage(lang)}
+                >
+                  {lang.toUpperCase()}
+                </a>
+              ))}
+            </div>
           </div>
         </footer>
       </Host>