+++ packages/components/src/components/bal-tag/bal-tag.tsx
@@ -23,6 +23,11 @@ export class Tag {
   @Prop() closable = false
 
   /**
+   * If `true` a light version of the color is displayed
+   */
+  @Prop() light = false
+
+  /**
    * @internal
    * Reduces the padding
    */
@@ -55,6 +60,7 @@ export class Tag {
       <Host
         class={{
           'tag': true,
+          'is-light': this.light,
           'is-dense': this.dense,
           'is-transparent': this.transparent,
           [this.colorCssClass]: true,
@@ -72,7 +78,7 @@ export class Tag {
           </bal-text>
           <bal-close
             style={{
-              display: this.closable ? 'inline-block' : 'none',
+              display: this.closable ? 'flex' : 'none',
             }}
             inverted
             background