+++ packages/components/src/components/bal-tag/stories/bal-tag.templates.ts
@@ -1,0 +1,24 @@
+export const colors = `
+<bal-tag color="" class="mr-1">Default</bal-tag>
+<bal-tag color="primary" class="mr-1">Primary</bal-tag>
+<bal-tag color="info" class="mr-1">Info</bal-tag>
+<bal-tag color="success" class="mr-1">Success</bal-tag>
+<bal-tag color="warning" class="mr-1">Warning</bal-tag>
+<bal-tag color="danger">Danger</bal-tag>
+<br>
+<br>
+<bal-tag color="purple" class="mr-1">Purple</bal-tag>
+<bal-tag color="red" class="mr-1">Red</bal-tag>
+<bal-tag color="yellow" class="mr-1">Yellow</bal-tag>
+<bal-tag color="green" class="mr-1">Green</bal-tag>
+
+`
+
+export const closable = `
+<bal-tag closable color="" class="mr-1">Default</bal-tag>
+<bal-tag closable color="primary" class="mr-1">Primary</bal-tag>
+<bal-tag closable color="info" class="mr-1">Info</bal-tag>
+<bal-tag closable color="success" class="mr-1">Success</bal-tag>
+<bal-tag closable color="warning" class="mr-1">Warning</bal-tag>
+<bal-tag closable color="danger">Danger</bal-tag>
+`